% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{addInteractionToVars}
\alias{addInteractionToVars}
\title{Add interactions of a specific order to a vector of variables}
\usage{
addInteractionToVars(order, vars)
}
\arguments{
\item{order}{integer: order of the interaction}

\item{vars}{character: variables}
}
\value{
Character vector
}
\description{
Add interactions of a specific order to a vector of variables
}
\details{
Interactions of variables with themselves (including polynomials
of themselves) are not included.
}
\examples{
BMSC:::addInteractionToVars(3, c("x1", "x2", "x3"))
}
