% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{addPowToVars}
\alias{addPowToVars}
\title{Add exponent to a vector of variables}
\usage{
addPowToVars(vars, power)
}
\arguments{
\item{vars}{character: variable names}

\item{power}{integer: exponent}
}
\value{
Vector of same length as \code{vars}
}
\description{
Remark: Since this function is to be used only within
\code{\link{createFormula}}, the validity of the input is not checked here
but in \code{\link{createFormula}}.
}
\examples{
BMSC:::addPowToVars(c("x1", "x2"), 2)
}
