% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateModel.R
\name{bestModel}
\alias{bestModel}
\title{Best model}
\usage{
bestModel(models, loos, thresholdSE, ic)
}
\arguments{
\item{models}{list of models fitted by \code{\link{constrSelEst}} function}

\item{loos}{list of model fits}

\item{thresholdSE}{numeric: How much standard errors in leave-one-out
prediction performance can the sparse model be worse than the best model}

\item{ic}{information criterion: one of "Loo", "AIC", "WAIC", "BIC", "logLik"}
}
\description{
Best model
}
