% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ConstrainedLinReg-class}
\alias{ConstrainedLinReg-class}
\alias{ConstrainedLinReg}
\title{S4 class for constrained linear regression models}
\description{
Inherits from \code{\link[rstan]{stanfit}}
}
\section{Slots}{

\describe{
\item{\code{formula}}{model formula (class formula)}

\item{\code{hasIntercept}}{logical: Does the model formula include an intercept?}

\item{\code{scaleCenter}}{numeric: location scale of betas}

\item{\code{scaleScale}}{numeric: scale scale of betas}
}}

