% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{createFormulaInternal}
\alias{createFormulaInternal}
\title{Create formula with interactions and polynomials if all checks in
\code{\link{createFormula}} have passed}
\usage{
createFormulaInternal(formula, allVars, maxExponent, interactionDepth,
  intercept)
}
\arguments{
\item{formula}{formula object}

\item{allVars}{object returned by \code{\link[base]{all.vars}}}

\item{maxExponent}{positive integer}

\item{interactionDepth}{positive integer}

\item{intercept}{boolean}
}
\description{
Create formula with interactions and polynomials if all checks in
\code{\link{createFormula}} have passed
}
