% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{extractVarname}
\alias{extractVarname}
\title{Extract variable name from polynomial expression}
\usage{
extractVarname(x)
}
\arguments{
\item{x}{Character: variables}
}
\description{
Extract variable name from polynomial expression
}
\examples{
BMSC:::extractVarname(c("x1",
"I(x2^2)"))
}
