% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_averaging.R
\name{get_model_weights}
\alias{get_model_weights}
\title{Compute model weights according to measure}
\usage{
get_model_weights(fits, measure = "Loo")
}
\arguments{
\item{fits}{List of model fits (return of getModelFits()-function)}

\item{measure}{string with type of measure}
}
\value{
Numeric vector of model weights
}
\description{
Computes model weights according to measure
}
