% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateModel.R
\name{getBestModel}
\alias{getBestModel}
\title{Get Best Model after Models Selection}
\usage{
getBestModel(models, thresholdSE = 1, plotModels = TRUE)
}
\arguments{
\item{models}{list of models fitted by \code{\link{constrSelEst}} function}

\item{thresholdSE}{numeric: How much standard errors in leave-one-out
prediction performance can the sparse model be worse than the best model}

\item{plotModels}{boolean: Plot models in leave-one-out evaluation plot TRUE/FALSE}
}
\value{
The best sparse model concerning leave-one-out performance within a
threshold
}
\description{
Get Best Model after Models Selection
}
