% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_show.R
\name{getBetaMatrix}
\alias{getBetaMatrix}
\title{Extract beta matrix from \code{\link{ConstrainedLinReg}} model}
\usage{
getBetaMatrix(model, hasIntercept)
}
\arguments{
\item{model}{model object: Model of class \code{\link{ConstrainedLinReg}}}

\item{hasIntercept}{logical: Does the model formula include an intercept?}
}
\value{
matrix of estimates
}
\description{
Extracts matrix of beta estimates
}
