% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateModel.R
\name{getModelFits}
\alias{getModelFits}
\title{Get Model Fits}
\usage{
getModelFits(models, y = NULL, newdata = NULL)
}
\arguments{
\item{models}{list of models fitted by \code{\link{constrSelEst}} function}

\item{y}{response variable}

\item{newdata}{data.frame containing all variables that appear in the model formula}
}
\description{
Get Model Fits
}
