% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handleMissingValues.R
\name{handleMissingData}
\alias{handleMissingData}
\title{Exclude rows with missing values}
\usage{
handleMissingData(data, formula, yUncertainty)
}
\arguments{
\item{data}{data.frame}

\item{formula}{formula object}

\item{yUncertainty}{numeric: vector}
}
\value{
A list with the elements "data" (data frame containing only the
relevant variables and complete rows) and "yUncertainty".
}
\description{
All rows with missing values on the variables from the model
formula are excluded. If all rows are excluded, an error occurs. If only some
of the rows are excluded, the number and percentage of excluded rows is
printed via a message. In addition, the corresponding positions from
the yUncertainty vector are excluded.
}
