% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{makeInteractions}
\alias{makeInteractions}
\title{Add all interactions up to a desired order}
\usage{
makeInteractions(vars, interactionDepth)
}
\arguments{
\item{vars}{character: variable names (potentially including polynomial
expressions)}

\item{interactionDepth}{integer: highest interaction order}
}
\value{
Character vector
}
\description{
Add all interactions up to a desired order
}
\details{
Interactions of variables with themselves (including polynomials
of themselves) are not included.
}
\examples{
BMSC:::makeInteractions(vars = c("x1", "x2",
"I(x1^2)", "I(x2^2)"), interactionDepth = 3)
}
