% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{plotModelFit}
\alias{plotModelFit}
\title{Plot errors of all models}
\usage{
plotModelFit(models, thresholdSE = 1, loos = NULL, markBestModel = TRUE)
}
\arguments{
\item{models}{List with models of class \code{\link{ConstrainedLinReg}}}

\item{thresholdSE}{numeric: Factor multiplied with standard error to obtain
ends of error bars}

\item{loos}{List with the model fit results for all models as returned by
\code{BMSC:::getLoo}. If not provided, they are
computed from the model list, which can take some time.}

\item{markBestModel}{boolean: highlight position of the best model in the model list}
}
\description{
This plot is automatically produced with the execution of
\code{\link{getBestModel}}.
}
