% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{plotModels}
\alias{plotModels}
\title{Plot model errors with errorbars}
\usage{
plotModels(datPlot, colours, thresholdSE)
}
\arguments{
\item{datPlot}{data.frame with prepared plot data}

\item{colours}{character: colour(s) for the points, bars and x-axis labels}

\item{thresholdSE}{numeric: Factor multiplied with standard error to obtain
ends of error bars}
}
\description{
Plot model errors with errorbars
}
