% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_predict.R
\docType{methods}
\name{predict,ConstrainedLinReg-method}
\alias{predict,ConstrainedLinReg-method}
\title{Compute predictions from constraint estimation model}
\usage{
\S4method{predict}{ConstrainedLinReg}(object, newdata)
}
\arguments{
\item{object}{Model of class \code{\link{ConstrainedLinReg}}}

\item{newdata}{data.frame containing all variables that appear in the model
formula}
}
\value{
Numeric vector of predictions. For observations with missing
values on the explanatory variables, a prediction of \code{NA} is returned.
}
\description{
Computes prediction from model of class
\code{\link{ConstrainedLinReg}} and a data.frame.
}
