% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{prepColorVec}
\alias{prepColorVec}
\title{Prepare colour vector}
\usage{
prepColorVec(posBestModel, length)
}
\arguments{
\item{posBestModel}{numeric: position of best Model}

\item{length}{numeric: Length of colour vector}
}
\value{
Vector of length \code{length}. It contains "black" expect for the
position provided in \code{posBestModel}, which is "chartreuse4" (green)
}
\description{
Prepare colour vector
}
