% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{prepPlotData}
\alias{prepPlotData}
\title{Prepare data to plot model fit}
\usage{
prepPlotData(loos, modelNames, thresholdSE)
}
\arguments{
\item{loos}{List with the model fit results for all models as returned by
\code{BMSC:::getLoo}. If not provided, they are
computed from the model list, which can take some time.}

\item{modelNames}{Names for the models in the same order as they appear in
\code{loos}}

\item{thresholdSE}{numeric: Factor multiplied with standard error to obtain
ends of error bars}
}
\value{
A data.frame with the columns \code{Estimate} (Estimate of the
looic), \code{SE}, \code{model}, \code{lower}, and \code{upper}
}
\description{
Prepare data to plot model fit
}
