% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_show.R
\name{print.ConstrainedLinReg}
\alias{print.ConstrainedLinReg}
\title{Print constraint estimation model}
\usage{
\method{print}{ConstrainedLinReg}(x, ...)
}
\arguments{
\item{x}{model object of class \code{\link{ConstrainedLinReg}}}

\item{...}{arguments passed from or to other methods}
}
\description{
Print constraint estimation model
}
