% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_show.R
\docType{methods}
\name{show,ConstrainedLinReg-method}
\alias{show,ConstrainedLinReg-method}
\title{Print constraint estimation model}
\usage{
\S4method{show}{ConstrainedLinReg}(object)
}
\arguments{
\item{object}{Model of class "MPIconstraintModel"}
}
\description{
Prints the model formula and estimates as well as sigma with
the corresponding 95% credibility intervals.
}
