% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{sortAndPaste}
\alias{sortAndPaste}
\title{Sort a vector and collapse elements together using ":"}
\usage{
sortAndPaste(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Sort a vector and collapse elements together using ":"
}
\examples{
BMSC:::sortAndPaste(c("var1", "var2"))
}
