% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFormula.R
\name{tryAsFormula}
\alias{tryAsFormula}
\title{Turn character vector into formula, return error if not possible}
\usage{
tryAsFormula(input)
}
\arguments{
\item{input}{character}
}
\value{
Formula or error
}
\description{
Turn character vector into formula, return error if not possible
}
