% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{plotModelFit}
\alias{plotModelFit}
\title{Plot errors of all models}
\usage{
plotModelFit(
  models,
  fits = NULL,
  thresholdSE = 1,
  markBestModel = TRUE,
  ic = "Loo",
  ylab = NULL,
  xlab = NULL,
  aSize = 12,
  lSize = 14,
  tAngle = 30
)
}
\arguments{
\item{models}{List with models of class \code{\link{ConstrainedLinReg}}}

\item{fits}{Optional list with model fit measures from getModelFits() - function}

\item{thresholdSE}{numeric: Factor multiplied with standard error to obtain
ends of error bars}

\item{markBestModel}{boolean: highlight position of the best model in the model list}

\item{ic}{information criterion: one of "Loo", "AIC", "WAIC", "BIC", "logLik"}

\item{ylab}{character: y title}

\item{xlab}{character: x title}

\item{aSize}{numeric: axis label size}

\item{lSize}{numeric: axis title label size}

\item{tAngle}{numeric: x axis text angle}
}
\description{
This plot is automatically produced with the execution of
\code{\link{getBestModel}}.
}
