% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{plotModels}
\alias{plotModels}
\title{Plot model errors with errorbars}
\usage{
plotModels(
  datPlot,
  colours,
  thresholdSE,
  ic,
  ylab = NULL,
  xlab = NULL,
  aSize = 12,
  lSize = 14,
  tAngle = 30
)
}
\arguments{
\item{datPlot}{data.frame with prepared plot data}

\item{colours}{character: colour(s) for the points, bars and x-axis labels}

\item{thresholdSE}{numeric: Factor multiplied with standard error to obtain
ends of error bars}

\item{ic}{information criterion: one of "Loo", "AIC", "WAIC", "BIC", "logLik"}

\item{ylab}{character: y title}

\item{xlab}{character: x title}

\item{aSize}{character: axis label size}

\item{lSize}{character: axis title label size}

\item{tAngle}{numeric: x axis text angle}
}
\description{
Plot model errors with errorbars
}
