% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelFit.R
\name{prepPlotData}
\alias{prepPlotData}
\title{Prepare data to plot model fit}
\usage{
prepPlotData(fits, modelNames, thresholdSE, ic = "Loo")
}
\arguments{
\item{fits}{Optional list with model fit measures from getModelFits() - function}

\item{modelNames}{Names for the models in the same order as they appear in
\code{Loos}}

\item{thresholdSE}{numeric: Factor multiplied with standard error to obtain
ends of error bars}

\item{ic}{information criterion: one of "Loo", "AIC", "WAIC", "BIC", "logLik"}
}
\value{
A data.frame with the columns \code{Estimate} (Estimate of the
Looic), \code{SE}, \code{model}, \code{lower}, and \code{upper}
}
\description{
Prepare data to plot model fit
}
